RIHAKB-484. Allow more properties for json_content entities
json_content internal processing as a special kind of property is
disabled when creating or updating records. When serialized or
deserialized, json_content is still treated as a raw JSON but now it
will be treated as a property of an entity allowing other entity
properties to be added to the response.